udpate the api for delete

+++ src/main/java/com/main/service/UserService.java
@@ -7,12 +7,15 @@ import lombok.NoArgsConstructor;
 import lombok.RequiredArgsConstructor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class UserService {
@@ -34,6 +37,11 @@ public class UserService {
     }
 
     public String deleteUserById(Long id) {
+        logger.info("Service request received to delete user by id: {}.", id);
+        Optional<UserEntity> user = this.userRepo.findById(id);
+        if(!user.isPresent()){
+            throw  new ResponseStatusException(HttpStatus.NOT_FOUND,"user not found with id : " + id);
+        }
         this.userRepo.deleteById(id);
         return "User Deleted successfully";
     }
